# services/notifications-service/internal/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "BearerAuth.Scopes"
)

// PostV1EmailSendJSONBody defines parameters for PostV1EmailSend.
type PostV1EmailSendJSONBody struct {
	// Message Cuerpo del mensaje (puede ser texto plano o HTML).
	Message string `json:"message"`

	// RecipientEmail Correo electrónico del destinatario.
	RecipientEmail openapi_types.Email `json:"recipientEmail"`

	// SenderEmail Correo electrónico del remitente. Debe estar configurado en el servidor SMTP.
	SenderEmail openapi_types.Email `json:"senderEmail"`

	// Subject Título o asunto del correo electrónico.
	Subject string `json:"subject"`
}

// PostV1EmailSendParams defines parameters for PostV1EmailSend.
type PostV1EmailSendParams struct {
	// XRequestID Identificador único para trazar la solicitud a través de múltiples servicios.
	XRequestID *string `json:"X-Request-ID,omitempty"`

	// XClientID Identificador del Operador que realiza la solicitud. La identidad principal se valida en el JWT.
	XClientID *string `json:"X-Client-ID,omitempty"`
}

// PostV1EmailSendJSONRequestBody defines body for PostV1EmailSend for application/json ContentType.
type PostV1EmailSendJSONRequestBody PostV1EmailSendJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Envía un correo electrónico con los detalles proporcionados.
	// (POST /v1/email/send)
	PostV1EmailSend(ctx echo.Context, params PostV1EmailSendParams) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// PostV1EmailSend converts echo context to params.
func (w *ServerInterfaceWrapper) PostV1EmailSend(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params PostV1EmailSendParams

	headers := ctx.Request().Header
	// ------------- Optional header parameter "X-Request-ID" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-Request-ID")]; found {
		var XRequestID string
		n := len(valueList)
		if n != 1 {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Expected one value for X-Request-ID, got %d", n))
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-Request-ID", valueList[0], &XRequestID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: false})
		if err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter X-Request-ID: %s", err))
		}

		params.XRequestID = &XRequestID
	}
	// ------------- Optional header parameter "X-Client-ID" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-Client-ID")]; found {
		var XClientID string
		n := len(valueList)
		if n != 1 {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Expected one value for X-Client-ID, got %d", n))
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-Client-ID", valueList[0], &XClientID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: false})
		if err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter X-Client-ID: %s", err))
		}

		params.XClientID = &XClientID
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostV1EmailSend(ctx, params)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/v1/email/send", wrapper.PostV1EmailSend)

}